Been working on some internal stuff. small bugfixes.
Added a new memorydetectionrouting that for the moment only is available as a menu-choise in memtest menu, PLEASE report results to me, especially if it finds different amount of memory than others.

And also false reports.. (rememember that you WILL need to do a autoconfig procerdure for autoconfig ram to be visible here)

also started to plan a new memorytestroutine that handles addresserrors etc.