Don't use stack until we know it works.
This code is executed way before RAM has been tested, so there's no
guarantee that we can push anything onto the stack.  The surrounding
code already makes this assumption, I suspect it is an oversight here
as the routine called returns via a jmp (a0) that we've already prepared
for.